feat: add inventory foundation and decimal quantity support

+++ app/schemas/stock_schema.py
@@ -1,11 +1,16 @@
 from datetime import datetime
+from decimal import Decimal
 
 from pydantic import BaseModel, ConfigDict, Field
 
 
 class StockIn(BaseModel):
     product_id: int = Field(..., gt=0)
-    quantity: int = Field(..., gt=0)
+
+    quantity: Decimal = Field(
+        ...,
+        gt=0,
+    )
 
     remark: str | None = Field(
         default=None,
@@ -15,7 +20,11 @@ class StockIn(BaseModel):
 
 class StockOut(BaseModel):
     product_id: int = Field(..., gt=0)
-    quantity: int = Field(..., gt=0)
+
+    quantity: Decimal = Field(
+        ...,
+        gt=0,
+    )
 
     remark: str | None = Field(
         default=None,
@@ -25,7 +34,11 @@ class StockOut(BaseModel):
 
 class StockAdjust(BaseModel):
     product_id: int = Field(..., gt=0)
-    new_quantity: int = Field(..., ge=0)
+
+    new_quantity: Decimal = Field(
+        ...,
+        ge=0,
+    )
 
     remark: str | None = Field(
         default=None,
@@ -36,16 +49,16 @@ class StockAdjust(BaseModel):
 class StockOperationResponse(BaseModel):
     product_id: int
     product_name: str
-    previous_stock: int
-    current_stock: int
-    difference: int
+    previous_stock: Decimal
+    current_stock: Decimal
+    difference: Decimal
 
 
 class StockTransactionResponse(BaseModel):
     id: int
     product_id: int
     transaction_type: str
-    quantity: int
+    quantity: Decimal
     remark: str | None = None
     created_at: datetime | None = None
